refactor: align lead engine client with configuration
